requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-RESTful==0.3.10
Flask-APScheduler==1.13.1
SQLAlchemy==2.0.41  
python-dotenv==1.0.1
Werkzeug==3.1.3
gunicorn==21.2.0  # For production deployment
pytest==8.4.0  # For testing  